On-The-Go Mode (2/2).
On-The-Go turns on the camera of your phone and streams it to an
overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the
content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

have fun!

Bring up from MM sources
Squash in alpha settings by LorDClockaN:
Modify alpha settings to use CustomSeekbarPreference

Signed-off-by: spezi77 <spezi77@gmx.com>